+++ R/http.R
@@ -1,4 +1,4 @@
-#' @title Generate a rate-limited API caller
+#' @title Generate a rate-limited Analytics Query Service API caller
 #' @description Generates a function which makes a request to
 #'   [wikimedia.org/api/rest_v1/](https://wikimedia.org/api/rest_v1/). This
 #'   generator is primarily intended for internal use, but can theoretically be
@@ -34,4 +34,46 @@ wx_query_api <- function(reqs_per_second = 200) {
     return(result)
   }
   return(limit_rate(http_get, rate(n = reqs_per_second, period = 1)))
+}
+
+#' @title Generate a rate-limited MediaWiki API caller
+#' @description Generates a function which makes a request to MediaWiki APIs.
+#'   This generator is primarily intended for internal use, but can
+#'   theoretically be used by others to write custom MediaWiki API queries. The
+#'   generated function is rate-limited at 100 queries per second.
+#' @param project The project for which to generate queryable MW API endpoint
+#' @return A rate-limited function with parameters:
+#' \describe{
+#'   \item{`query`}{query after endpoint; e.g. `action=query&format=json&titles=Main%20page`}
+#' }
+#' @seealso
+#' - [MediaWiki API](https://www.mediawiki.org/wiki/API:Main_page)
+#' @import ratelimitr
+wx_mediawiki_api <- function(project) {
+  if (!grepl("\\.org$", project)) {
+    active_wikis <- utils::read.csv(system.file("extdata/wikis.csv", package = "waxer"))
+    if (!project %in% active_wikis$project) {
+      stop()
+    } else {
+      base_url <- paste0("https://", active_wikis$url[active_wikis$project == project])
+    }
+  } else {
+    base_url <- paste0("https://", project) # e.g. mediawiki.org
+  }
+  http_get <- function(query) {
+    response <- httr::GET(
+      base_url, path = "w/api.php", query = query,
+      httr::user_agent("R wrapper for Wikimedia AQS API (https://github.com/bearloga/waxer)"),
+      httr::accept_json()
+    )
+    httr::stop_for_status(response)
+    if (httr::http_type(response) == "application/json") {
+      result <- jsonlite::fromJSON(
+        httr::content(response, as = "text", encoding = "UTF-8"),
+        simplifyVector = FALSE
+      )
+    }
+    return(result)
+  }
+  return(limit_rate(http_get, rate(n = 100, period = 1)))
 }